requirements.txt
pydantic
python-dotenv
openai
requests
pypdf
slowapi
